+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file. This projec
 -   add new package @comet/admin-icons
 -   add onAfterSubmit to FinalForm
 -   add useStoredState() hook
+-   Added a new FinalFormRangeInput Component
 
 ### Incompatible Changes